Update index.html
Updated <title>.
Added Google Fonts link for "Source Sans Pro" and "Source Serif Pro".
Restructured the <body> with div.container-main, header.date-display, and main.quote-jumbotron.
Added a placeholder <p id="gregorianDate">.
Changed the ID for the main quote text display to h1#quote-text.
Changed the ID for the author/source display to h2#quote-author-source.
The loading message for the quote is updated.
Script tags remain the same and in the same order.